Update PoolDashboard, Dashboard components call useGetPoolsQuery (PoolListApi) within Dashboard component
